Add files via upload
this is the solution of   let us c   chapter 2

C instructions are the commands in the program that instructs the compiler to do certain action. Basically it gives the instruction to the compiler on how to achieve the goal of the program. For example, when we say add two numbers; C compiler will not understand how to do it. But if we write the same in terms of code / command using arithmetic operator ‘+’ and numbers or variables, it gives meaningful information to the compiler on how to get the sum.

Note:- I am not an expert. I am Engineering student. So my answer may have error or mistake